[BF-50] Fix : suppression comments par users via url

+++ src/Router/CommentRouter.php
@@ -20,6 +20,11 @@ class CommentRouter extends AbstractRouter {
     }
 
     public function delete(){
+        if(!$this->previous instanceof AdminRouter){
+            $this->_403();
+            return;
+        }
+
         $controller = new \Controller\CommentController();
         $controller->deleteComment(
             $this->url[0],
